refactor testcontainers configuration implementation to fixtures
