+++ integrationtesting/fixtures/src/main/java/org/opendcs/fixtures/configuration/ConfigurationProvider.java
@@ -1,4 +1,4 @@
-package org.opendcs.spi.configuration;
+package org.opendcs.fixtures.configuration;
 
 import java.io.File;
