+++ integrationtesting/fixtures/src/main/java/org/opendcs/fixtures/configuration/UserPropertiesBuilder.java
@@ -1,4 +1,4 @@
-package org.opendcs.fixtures;
+package org.opendcs.fixtures.configuration;
 
 import java.io.IOException;
 import java.io.OutputStream;
