+++ integrationtesting/fixtures/src/main/java/org/opendcs/fixtures/configuration/xml/XmlConfiguration.java
@@ -1,4 +1,4 @@
-package org.opendcs.fixtures.configurations.xml;
+package org.opendcs.fixtures.configuration.xml;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -10,15 +10,12 @@ import java.util.logging.Logger;
 import org.apache.commons.io.FileUtils;
 import org.opendcs.database.DatabaseService;
 import org.opendcs.database.api.OpenDcsDatabase;
-import org.opendcs.fixtures.UserPropertiesBuilder;
-import org.opendcs.spi.configuration.Configuration;
+import org.opendcs.fixtures.configuration.Configuration;
 
 import decodes.db.Database;
-import decodes.db.DatabaseIO;
 import decodes.launcher.Profile;
-import decodes.tsdb.TimeSeriesDb;
 import decodes.util.DecodesSettings;
-import ilex.util.Pair;
+import org.opendcs.fixtures.configuration.UserPropertiesBuilder;
 import uk.org.webcompere.systemstubs.environment.EnvironmentVariables;
 import uk.org.webcompere.systemstubs.properties.SystemProperties;
 import uk.org.webcompere.systemstubs.security.SystemExit;
