+++ integrationtesting/opendcs-tests/build.gradle
@@ -9,14 +9,8 @@ dependencies {
     testCompileOnly project(":opendcs")
     testImplementation project(":testing:fixtures")
     testImplementation enforcedPlatform(libs.junit.bom)
-    testImplementation(libs.testcontainers.core)
-    testImplementation(libs.testcontainers.postgresql)
-    testImplementation(libs.hec.testcontainers.cwms)
-    testImplementation(libs.bundles.jdbi)
     testImplementation(libs.bundles.junit)
     testImplementation(libs.bundles.junit.platform)
-    testImplementation(libs.commons.io)
-    testImplementation(libs.apache.derby)
     testImplementation(libs.fasterxml.jackson.databind)
     testImplementation(libs.webcompere.system.stubs.jupiter)
 }
