+++ integrationtesting/opendcs-tests/src/test/java/org/opendcs/fixtures/annotations/EnableIfAppsSupportedCondition.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.extension.ConditionEvaluationResult;
 import org.junit.jupiter.api.extension.ExecutionCondition;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.platform.commons.support.AnnotationSupport;
-import org.opendcs.spi.configuration.Configuration;
+import org.opendcs.fixtures.configuration.Configuration;
 
 /**
  * Simple check to help avoid methods getting called in implementations
