+++ integrationtesting/opendcs-tests/src/test/java/org/opendcs/regression_tests/DecodesTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.params.provider.CsvSource;
 import org.opendcs.fixtures.AppTestBase;
 import org.opendcs.fixtures.annotations.DecodesConfigurationRequired;
 import org.opendcs.fixtures.helpers.TestResources;
-import org.opendcs.spi.configuration.Configuration;
+import org.opendcs.fixtures.configuration.Configuration;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
